requirements.txt
astral==0.9
boto==2.34.0
crypto==1.4.1
Django==1.8.8
django-autoslug==1.9.3
django-cities-light==3.2.0
django-contrib-comments==1.5
django-debug-toolbar==1.4
django-markdown-deux==1.0.5
django-mptt==0.7.1
django-nocaptcha-recaptcha==0.0.18
django-ses==0.6.0
django-session-security==2.4.0
django-storages==1.1.8
django-tagging==0.3.3
django-paypal==0.2.5
Flutterwave==1.0.5
Markdown==2.6.2
markdown2==2.3.1
MySQL-python==1.2.5
Naked==0.1.31
Pillow==2.6.1
progressbar2==3.9.3
pycrypto==2.6.1
pyparsing==2.0.3
python-utils==2.0.0
pytz==2014.7
PyYAML==3.12
reportlab==3.2.0
requests==2.11.1
shellescape==3.4.1
six==1.10.0
sqlparse==0.1.19
Unidecode==0.4.19
django-localflavor==1.3
django-model-utils==2.6
pycountry
py-moneyed
forex-python
paystackapi==1.2.6